refactor(workspace): extract the artifact viewer's shared primitives (§13.1 stage one)
The primitives every artifact family is built from now live in
`src/components/workspace/artifactViewers/`, ahead of moving the families
themselves. Extracting these first is what stops the family moves from
producing import cycles.

- `CollapsibleSection.tsx`, `TextCopyButton.tsx`, `MetadataCard.tsx`
  (`MetadataCard` + `ArtifactListSection`), `artifactTimestamp.ts`,
  `rawContentSources.ts` (the raw source/variant types and selection logic)
  and `WithRawTab.tsx` (`WithRawTab` + `RawAttemptVariantSelector`).
- `useActiveRawVariant` in `rawContentSources.ts` replaces the variant-snapping
  effect that `FinalPrdDraftView` and `RelevantFilesScanView` each wrote out,
  identically apart from the id they fall back to. The fallback order is
  preserved exactly: selected-and-enabled, then the first validated variant,
  then the first enabled one, then the first variant at all.

`ArtifactContentViewer.tsx` goes from 7,379 to 6,862 lines and still exports
`CollapsibleSection`, `TextCopyButton`, `WithRawTab` and `CopyButton`, so
`PhaseArtifactsPanel` and the artifact tests are untouched.

Six components imported `CollapsibleSection` from the viewer and nothing else,
which dragged the whole module and its 2,275-line parser neighbour into their
bundle. `PrdDocumentView`, `CoverageApprovalWarning`, `PrdApprovalEditor`,
`InterviewApprovalAnswerEditor`, `InterviewDocumentView` and `ManualQAView`
now import it from the new module directly. This is a departure from the
section text, which asked for no import-site changes at all; without it the
re-export keeps the old graph and the split buys nothing at runtime. Agreed
with the owner before starting.

Departure, §13.1's raw-statistics item: the two blocks it names at
`:1405-1407,1507-1510` are not reimplemented raw-text statistics. They are
diff counts — Modified / Added / Removed / Replaced — and `RawDisplayStats`
derives lines, characters and tokens from a string, so it cannot express them.
What the three places actually share is the pill markup, so that is what was
extracted: `StatPillRow.tsx`, used by both diff views and by `RawDisplayStats`
itself. The class strings are unchanged, so nothing renders differently.

Impact: no visual, class-name or copy change; the 551 workspace component
tests pass unchanged. Bundling improves for the six repointed components.
