+++ src/components/workspace/PrdDocumentView.tsx
@@ -11,7 +11,7 @@ import {
   getPrdTechnicalRequirementAnchorId,
   getPrdTechnicalRequirementsAnchorId,
 } from '@/lib/prdDocument'
-import { CollapsibleSection } from './ArtifactContentViewer'
+import { CollapsibleSection } from './artifactViewers/CollapsibleSection'
 
 function MetaPill({ children, className }: { children: React.ReactNode; className?: string }) {
   return (
